Update and rename functies.py to codereview_kai.py
Niet alles onder elkaar zetten, lege regels gebruiken.
Comments gebruiken voor begrijpelijkheid
Regel 19, één for loop gebruiken i.p.v. twee.
