--- functies.py
@@ -1,28 +1,0 @@
-def combinatielijst(elementenlijst, lengte):
-    import itertools as it
-    allecombinaties = []
-
-
-    for j in it.product((elementenlijst), repeat=lengte):
-        allecombinaties.append(j)
-    return allecombinaties
-
-
-
-
-
-def pinchecker(Gokken, PCcode):
-    Zwart = 0
-    Wit = 0
-    Algebruikt = []
-    for i in range(4):
-        if Gokken[i] == PCcode[i]:
-            Zwart += 1
-            Algebruikt.append(Gokken[i])
-
-    for j in range(4):
-        if Gokken[j] in PCcode and Gokken[j] != PCcode[j] and Gokken[j] not in Algebruikt:
-            Wit += 1
-
-    return Wit, Zwart
-
